// src/generated/models/Geek_projecrtsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface Geek_projecrts {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: Date;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  geek_description?: string;
  geek_end?: Date;
  geek_id?: string;
  geek_manager?: string;
  geek_name?: string;
  geek_projecrtid: string;
  geek_start?: Date;
  geek_status?: string;
  geek_statusname?: string;
  importsequencenumber?: number;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: Date;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: Date;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: string;
  statecodename?: string;
  statuscode?: string;
  statuscodename?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  versionnumber?: number;
}